feat: stateless transformer works :))
