+++ src/data/exchange/binance/mod.rs
@@ -12,11 +12,12 @@ use crate::data::{
         book::OrderBookL2,
         subs::{ExchangeId, StreamType},
         trade::Trades,
-    },
-    protocols::ws::ws_client::WsMessage,
-    transformer::stateless_transformer::StatelessTransformer,
+    }, protocols::ws::WsMessage, transformer::stateless_transformer::StatelessTransformer
 };
 
+/*----- */
+// Binance connector
+/*----- */
 #[derive(Debug, Default, Eq, PartialEq, Hash)]
 pub struct BinanceSpot;
 
@@ -61,12 +62,13 @@ impl Connector for BinanceSpot {
     }
 }
 
+/*----- */
+// Stream selector
+/*----- */
 impl StreamSelector<BinanceSpot, OrderBookL2> for BinanceSpot {
-    type Stream = BinanceBook;
     type StreamTransformer = StatelessTransformer<BinanceBook, OrderBookL2>;
 }
 
 impl StreamSelector<BinanceSpot, Trades> for BinanceSpot {
-    type Stream = BinanceTrade;
     type StreamTransformer = StatelessTransformer<BinanceTrade, Trades>;
 }
