+++ src/data/exchange/mod.rs
@@ -1,17 +1,15 @@
 pub mod binance;
 pub mod poloniex;
 
-use std::fmt::Debug;
-
 use serde::de::DeserializeOwned;
 
 use super::{
     models::{
-        event::MarketEvent,
         subs::{ExchangeId, Instrument},
         SubKind,
     },
-    protocols::ws::ws_client::{PingInterval, WsMessage}, transformer::Transformer,
+    protocols::ws::{PingInterval, WsMessage},
+    transformer::Transformer,
 };
 
 /*----- */
@@ -20,8 +18,6 @@ use super::{
 pub trait Connector {
     type ExchangeId;
     type SubscriptionResponse: DeserializeOwned;
-    // type Channel: AsRef<str>;
-    // type Market: AsRef<str>;
 
     const ID: ExchangeId;
 
@@ -43,9 +39,7 @@ pub trait StreamSelector<Exchange, StreamKind>
 where
     Exchange: Connector,
     StreamKind: SubKind,
-    
 {
-    type Stream: DeserializeOwned + Into<MarketEvent<StreamKind::Event>> + Debug + Send;
     type StreamTransformer: Transformer + Default + Send;
 }
