+++ src/data/exchange/poloniex/mod.rs
@@ -10,9 +10,12 @@ use super::{Connector, Instrument, StreamSelector};
 use crate::data::models::book::OrderBookL2;
 use crate::data::models::subs::{ExchangeId, StreamType};
 use crate::data::models::trade::Trades;
-use crate::data::protocols::ws::ws_client::{PingInterval, WsMessage};
+use crate::data::protocols::ws::{PingInterval, WsMessage};
 use crate::data::transformer::stateless_transformer::StatelessTransformer;
 
+/*----- */
+// Poloniex connector
+/*----- */
 #[derive(Debug, Default, Eq, PartialEq, Hash)]
 pub struct PoloniexSpot;
 
@@ -68,12 +71,13 @@ impl Connector for PoloniexSpot {
     }
 }
 
+/*----- */
+// Stream selector
+/*----- */
 impl StreamSelector<PoloniexSpot, OrderBookL2> for PoloniexSpot {
-    type Stream = PoloniexBook;
     type StreamTransformer = StatelessTransformer<PoloniexBook, OrderBookL2>;
 }
 
 impl StreamSelector<PoloniexSpot, Trades> for PoloniexSpot {
-    type Stream = PoloniexTrade;
     type StreamTransformer = StatelessTransformer<PoloniexTrade, Trades>;
 }
