+++ src/data/protocols/ws/mod.rs
@@ -1,225 +1,172 @@
-pub mod ws_client;
+pub mod ws_parser;
 
-use std::fmt::Debug;
-
-use futures::{Stream, StreamExt};
-use serde::de::DeserializeOwned;
-use tokio::{
-    sync::mpsc::UnboundedSender,
-    time::{sleep, Duration},
+use futures::{
+    stream::{SplitSink, SplitStream},
+    SinkExt, Stream, StreamExt,
 };
-use tokio_tungstenite::tungstenite::{
-    error::ProtocolError,
-    protocol::{frame::Frame, CloseFrame},
+use pin_project::pin_project;
+use serde_json::Value;
+use std::{
+    collections::VecDeque,
+    marker::PhantomData,
+    pin::Pin,
+    task::{Context, Poll},
 };
-use ws_client::{ExchangeStream, WebSocket, WebSocketClient, WsError, WsMessage};
+use tokio::{net::TcpStream, time::sleep, time::Duration};
+use tokio_tungstenite::{connect_async, MaybeTlsStream, WebSocketStream};
+use ws_parser::{StreamParser, WebSocketParser};
 
 use crate::{
     data::{
         exchange::{Connector, StreamSelector},
-        models::{event::MarketEvent, subs::Subscription, SubKind},
+        models::{subs::Instrument, SubKind},
         transformer::Transformer,
     },
     error::SocketError,
 };
 
-pub const START_RECONNECTION_BACKOFF_MS: u64 = 125;
+pub type WsMessage = tokio_tungstenite::tungstenite::Message;
+pub type WsError = tokio_tungstenite::tungstenite::Error;
+pub type WebSocket = tokio_tungstenite::WebSocketStream<MaybeTlsStream<TcpStream>>;
+pub type WsRead = SplitStream<WebSocketStream<MaybeTlsStream<TcpStream>>>;
+pub type WsWrite = SplitSink<WebSocketStream<MaybeTlsStream<TcpStream>>, WsMessage>;
+pub type JoinHandle = tokio::task::JoinHandle<()>;
 
-pub async fn connect<Exchange, StreamKind>(
-    exchange_sub: Vec<Subscription<Exchange, StreamKind>>,
-    exchange_tx: UnboundedSender<MarketEvent<StreamKind::Event>>,
-    // exchange_tx: UnboundedSender<<Exchange::StreamTransformer as Transformer>::Output>,
-) -> SocketError
+/*----- */
+// Models
+/*----- */
+#[derive(Clone, Debug)]
+pub struct PingInterval {
+    pub time: u64,
+    pub message: Value,
+}
+
+#[derive(Debug)]
+#[pin_project]
+pub struct ExchangeStream<StreamTransformer>
 where
-    Exchange: Connector + Send + StreamSelector<Exchange, StreamKind>,
-    StreamKind: SubKind,
-    // StreamKind::Event: Into<<Exchange::StreamTransformer as Transformer>::Output>,
-    <Exchange::StreamTransformer as Transformer>::Output: Into<MarketEvent<StreamKind::Event>>, // StreamKind::Event: Into<<Exchange::StreamTransformer as Transformer>::Output>
-                                                                                                // ExchangeStream<Exchange, StreamKind, Exchange::StreamTransformer>: Stream,
+    StreamTransformer: Transformer,
 {
-    let mut _connection_attempt: u32 = 0;
-    let mut _backoff_ms: u64 = START_RECONNECTION_BACKOFF_MS;
+    #[pin]
+    pub ws_read: WsRead,
+    pub transformer: StreamTransformer,
+    pub tasks: Vec<JoinHandle>,
+    pub buffer: VecDeque<Result<StreamTransformer::Output, StreamTransformer::Error>>,
+}
 
-    // I DONT LIKE THIS - MAKE IT CLEANER
-    let subs = exchange_sub
-        .into_iter()
-        .map(|s| s.instrument)
-        .collect::<Vec<_>>();
+impl<StreamTransformer> ExchangeStream<StreamTransformer>
+where
+    StreamTransformer: Transformer,
+{
+    pub fn new(stream: WsRead, transformer: StreamTransformer, tasks: Vec<JoinHandle>) -> Self {
+        Self {
+            ws_read: stream,
+            transformer,
+            tasks,
+            buffer: VecDeque::with_capacity(6),
+        }
+    }
 
-    loop {
-        _connection_attempt += 1;
-        _backoff_ms *= 2;
+    pub fn cancel_running_tasks(&self) {
+        self.tasks.iter().for_each(|task| {
+            task.abort();
+        })
+    }
+}
 
-        // Attempt to connect to the stream
-        let mut stream =
-            match WebSocketClient::<Exchange, StreamKind>::create_websocket(&subs).await {
-                Ok(stream) => {
-                    _connection_attempt = 0;
-                    _backoff_ms = START_RECONNECTION_BACKOFF_MS;
-                    stream
-                }
-                Err(error) => {
-                    if _connection_attempt == 1 {
-                        return SocketError::Subscribe(format!(
-                            "Subscription failed on first attempt: {}",
-                            error
-                        ));
-                    } else {
-                        continue;
-                    }
-                }
+impl<StreamTransformer> Stream for ExchangeStream<StreamTransformer>
+where
+    StreamTransformer: Transformer,
+    // StreamTransformer::Error: From<SocketError>,
+{
+    type Item = Result<StreamTransformer::Output, StreamTransformer::Error>;
+
+    fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
+        loop {
+            // Flush Self::Item buffer if it is not currently empty
+            if let Some(output) = self.buffer.pop_front() {
+                return Poll::Ready(Some(output));
+            }
+
+            // Poll inner `Stream` for next the next input protocol message
+            let input = match self.as_mut().project().ws_read.poll_next(cx) {
+                Poll::Ready(Some(input)) => input,
+                Poll::Ready(None) => return Poll::Ready(None),
+                Poll::Pending => return Poll::Pending,
             };
 
-        // Validate subscriptions
-        if let Some(Ok(WsMessage::Text(message))) = &stream.ws_read.next().await {
-            let subscription_sucess = Exchange::validate_subscription(message.to_owned(), &subs);
+            // Parse input protocol message into `ExchangeMessage`
+            let exchange_message =
+                match <WebSocketParser as StreamParser>::parse::<StreamTransformer::Input>(input) {
+                    // `StreamParser` successfully deserialised `ExchangeMessage`
+                    Some(Ok(exchange_message)) => exchange_message,
 
-            if !subscription_sucess {
-                break SocketError::Subscribe(String::from("Subscription failed"));
-            }
-        };
+                    // If `StreamParser` returns an Err pass it downstream
+                    // Some(Err(err)) => return Poll::Ready(Some(Err(err.into()))),
+                    Some(Err(_err)) => continue,
 
-        // Read from stream and send via channel, but if error occurs, attempt reconnection
-        while let Some(market_event) = stream.next().await {
-            match market_event {
-                Ok(market_event) => {
-                    println!("--- polled next ---");
-                    println!("{:#?}", market_event);
-                    exchange_tx
-                        .send(market_event.into())
-                        .expect("failed to send message");
-                }
-                Err(_error) => {
-                    // SHOULD CHANGE THIS TO SEE WHAT ERRORS ACTUAL MEANS INSTEAD OF BREAKING
-                    stream.cancel_running_tasks();
-                    break;
-                } // ADD Error for if sequence is broken then have to restart
-            }
-        }
+                    // If `StreamParser` returns None it's a safe-to-skip message
+                    None => continue,
+                };
 
-        // Wait a certain ms before trying to reconnect
-        sleep(Duration::from_millis(_backoff_ms)).await;
-    }
-}
-
-/*----- */
-// WebSocket message parser
-/*----- */
-pub trait StreamParser {
-    type Stream: Stream;
-    type Message;
-    type Error;
-
-    fn parse<Output>(
-        input: Result<Self::Message, Self::Error>,
-    ) -> Option<Result<Output, SocketError>>
-    where
-        Output: DeserializeOwned;
-}
-pub struct WebSocketParser;
-
-impl StreamParser for WebSocketParser {
-    type Stream = WebSocket;
-    type Message = WsMessage;
-    type Error = WsError;
-
-    fn parse<Output>(
-        input: Result<Self::Message, Self::Error>,
-    ) -> Option<Result<Output, SocketError>>
-    where
-        Output: DeserializeOwned,
-    {
-        match input {
-            Ok(ws_message) => match ws_message {
-                WsMessage::Text(text) => process_text(text),
-                WsMessage::Binary(binary) => process_binary(binary),
-                WsMessage::Ping(ping) => process_ping(ping),
-                WsMessage::Pong(pong) => process_pong(pong),
-                WsMessage::Close(close_frame) => process_close_frame(close_frame),
-                WsMessage::Frame(frame) => process_frame(frame),
-            },
-            Err(ws_err) => Some(Err(SocketError::WebSocketError(ws_err))),
+            let transformed_message = self.transformer.transform(exchange_message);
+            self.buffer.push_back(Ok(transformed_message))
         }
     }
 }
 
-pub fn parse<Output>(input: WsMessage) -> Option<Result<Output, SocketError>>
+/*----- */
+// WebSocket client
+/*----- */
+#[derive(Debug)]
+pub struct WebSocketClient<Exchange, StreamKind> {
+    pub phantom_markers: PhantomData<(Exchange, StreamKind)>,
+}
+
+impl<Exchange, StreamKind> WebSocketClient<Exchange, StreamKind>
 where
-    Output: DeserializeOwned,
+    Exchange: Connector + StreamSelector<Exchange, StreamKind> + Send,
+    StreamKind: SubKind,
 {
-    match input {
-        WsMessage::Text(text) => process_text(text),
-        WsMessage::Binary(binary) => process_binary(binary),
-        WsMessage::Ping(ping) => process_ping(ping),
-        WsMessage::Pong(pong) => process_pong(pong),
-        WsMessage::Close(close_frame) => process_close_frame(close_frame),
-        WsMessage::Frame(frame) => process_frame(frame),
+    pub async fn create_websocket(
+        subs: &[Instrument],
+    ) -> Result<ExchangeStream<Exchange::StreamTransformer>, SocketError> {
+        // Make connection
+        let mut tasks = Vec::new();
+        let ws = connect_async(Exchange::url().clone())
+            .await
+            .map(|(ws, _)| ws)
+            .map_err(SocketError::WebSocketError);
+
+        // Split WS and make channels
+        let (mut ws_write, ws_read) = ws?.split();
+
+        // Handle subscription
+        if let Some(subcription) = Exchange::requests(subs).clone() {
+            ws_write
+                .send(subcription)
+                .await
+                .expect("Failed to send subscription")
+        }
+
+        // Spawn custom ping handle (application level ping)
+        if let Some(ping_interval) = Exchange::ping_interval().clone() {
+            let ping_handler = tokio::spawn(schedule_pings_to_exchange(ws_write, ping_interval));
+            tasks.push(ping_handler);
+        }
+
+        let transformer = Exchange::StreamTransformer::default();
+
+        Ok(ExchangeStream::new(ws_read, transformer, tasks))
     }
 }
 
-pub fn process_text<ExchangeMessage>(
-    payload: String,
-) -> Option<Result<ExchangeMessage, SocketError>>
-where
-    ExchangeMessage: DeserializeOwned,
-{
-    Some(
-        serde_json::from_str::<ExchangeMessage>(&payload)
-            .map_err(|error| SocketError::Deserialise { error, payload }),
-    )
-}
-
-pub fn process_binary<ExchangeMessage>(
-    payload: Vec<u8>,
-) -> Option<Result<ExchangeMessage, SocketError>>
-where
-    ExchangeMessage: DeserializeOwned,
-{
-    Some(
-        serde_json::from_slice::<ExchangeMessage>(&payload).map_err(|error| {
-            SocketError::Deserialise {
-                error,
-                payload: String::from_utf8(payload).unwrap_or_else(|x| x.to_string()),
-            }
-        }),
-    )
-}
-
-pub fn process_ping<ExchangeMessage>(
-    ping: Vec<u8>,
-) -> Option<Result<ExchangeMessage, SocketError>> {
-    format!("{:#?}", ping);
-    None
-}
-
-pub fn process_pong<ExchangeMessage>(
-    pong: Vec<u8>,
-) -> Option<Result<ExchangeMessage, SocketError>> {
-    format!("{:#?}", pong);
-    None
-}
-
-pub fn process_close_frame<ExchangeMessage>(
-    close_frame: Option<CloseFrame<'_>>,
-) -> Option<Result<ExchangeMessage, SocketError>> {
-    let close_frame = format!("{:?}", close_frame);
-    Some(Err(SocketError::Terminated(close_frame)))
-}
-
-pub fn process_frame<ExchangeMessage>(
-    frame: Frame,
-) -> Option<Result<ExchangeMessage, SocketError>> {
-    format!("{:?}", frame);
-    None
-}
-
-pub fn is_websocket_disconnected(error: &WsError) -> bool {
-    matches!(
-        error,
-        WsError::ConnectionClosed
-            | WsError::AlreadyClosed
-            | WsError::Io(_)
-            | WsError::Protocol(ProtocolError::SendAfterClosing)
-            | WsError::Protocol(ProtocolError::ResetWithoutClosingHandshake) // | WsError::Protocol(_)
-    )
+pub async fn schedule_pings_to_exchange(mut ws_write: WsWrite, ping_interval: PingInterval) {
+    loop {
+        sleep(Duration::from_secs(ping_interval.time)).await;
+        ws_write
+            .send(WsMessage::Text(ping_interval.message.to_string()))
+            .await
+            .expect("Failed to send ping to ws");
+    }
 }
